+++ packages/playwright-test/src/DEPS.list
@@ -3,3 +3,4 @@
 matchers/
 reporters/
 third_party/
+plugins/webServerPlugin.ts
