+++ packages/playwright-test/src/loader.ts
@@ -30,6 +30,7 @@ import type { Reporter } from '../types/testReporter';
 import { builtInReporters } from './runner';
 import { isRegExp } from 'playwright-core/lib/utils';
 import { serializeError } from './util';
+import { webServer } from './plugins/webServerPlugin';
 
 // To allow multiple loaders in the same process without clearing require cache,
 // we make these maps global.
@@ -76,6 +77,11 @@ export class Loader {
   }
 
   private _processConfigObject(config: Config, configDir: string) {
+    if (config.webServer) {
+      config.plugins = config.plugins || [];
+      config.plugins.push(webServer(config.webServer));
+    }
+
     for (const plugin of config.plugins || [])
       plugin.configure?.(config, configDir);
