+++ tests/playwright-test/web-server.spec.ts
@@ -426,3 +426,165 @@ test(`should suport self signed certificate`, async ({ runInlineTest, httpsServe
   });
   expect(result.exitCode).toBe(0);
 });
+
+test('should create multiple servers', async ({ runInlineTest }, { workerIndex }) => {
+  const port = workerIndex + 10500;
+  const result = await runInlineTest({
+    'test.spec.ts': `
+        const { test } = pwt;
+        test('connect to the server', async ({baseURL, page}) => {
+          await page.goto('http://localhost:${port}/port');
+          await page.locator('text=${port}');
+
+          await page.goto('http://localhost:${port + 1}/port');
+          await page.locator('text=${port + 1}');
+        });
+      `,
+    'playwright.config.ts': `
+        import { webServer } from '@playwright/test/lib/plugins';
+        module.exports = {
+          plugins: [
+            webServer({
+                command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port}',
+                port: ${port},
+            }),
+            webServer({
+                command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port + 1}',
+                port: ${port + 1},
+            }),
+           ],
+           globalSetup: 'globalSetup.ts',
+           globalTeardown: 'globalTeardown.ts',
+          };
+        `,
+    'globalSetup.ts': `
+          module.exports = async () => {
+            const http = require("http");
+            const response = await new Promise(resolve => {
+              const request = http.request("http://localhost:${port}/hello", resolve);
+              request.end();
+            })
+            console.log('globalSetup-status-'+response.statusCode)
+            return async () => {
+              const response = await new Promise(resolve => {
+                const request = http.request("http://localhost:${port}/hello", resolve);
+                request.end();
+              })
+              console.log('globalSetup-teardown-status-'+response.statusCode)
+            };
+          };
+        `,
+    'globalTeardown.ts': `
+          module.exports = async () => {
+            const http = require("http");
+            const response = await new Promise(resolve => {
+              const request = http.request("http://localhost:${port}/hello", resolve);
+              request.end();
+            })
+            console.log('globalTeardown-status-'+response.statusCode)
+          };
+        `,
+  }, undefined, { DEBUG: 'pw:webserver' });
+  expect(result.exitCode).toBe(0);
+  expect(result.passed).toBe(1);
+  expect(result.output).toContain('[WebServer] listening');
+  expect(result.output).toContain('[WebServer] error from server');
+  expect(result.output).toContain('passed');
+
+  const expectedLogMessages = ['globalSetup-status-200', 'globalSetup-teardown-status', 'globalTeardown-status-200'];
+  const actualLogMessages = expectedLogMessages.map(log => ({
+    log,
+    index: result.output.indexOf(log),
+  })).sort((a, b) => a.index - b.index).filter(l => l.index !== -1).map(l => l.log);
+  expect(actualLogMessages).toStrictEqual(expectedLogMessages);
+});
+
+test.describe('baseURL with plugins', () => {
+  test('first plugin is selected as baseURL', async ({ runInlineTest }, { workerIndex }) => {
+    const port = workerIndex + 10500;
+    const result = await runInlineTest({
+      'test.spec.ts': `
+          const { test } = pwt;
+          test('connect to the server', async ({baseURL, page}) => {
+            expect(baseURL).toBe('http://localhost:${port}');
+          });
+        `,
+      'playwright.config.ts': `
+          import { webServer } from '@playwright/test/lib/plugins';
+          module.exports = {
+            plugins: [
+              webServer({
+                  command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port}',
+                  port: ${port},
+              }),
+              webServer({
+                  command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port + 1}',
+                  port: ${port + 1},
+              }),
+             ],
+            };
+          `,
+    }, undefined, { DEBUG: 'pw:webserver' });
+    expect(result.exitCode).toBe(0);
+    expect(result.passed).toBe(1);
+  });
+
+  test('setBaseURL=false works', async ({ runInlineTest }, { workerIndex }) => {
+    const port = workerIndex + 10500;
+    const result = await runInlineTest({
+      'test.spec.ts': `
+          const { test } = pwt;
+          test('connect to the server', async ({baseURL, page}) => {
+            expect(baseURL).toBe('http://localhost:${port + 1}');
+          });
+        `,
+      'playwright.config.ts': `
+          import { webServer } from '@playwright/test/lib/plugins';
+          module.exports = {
+            plugins: [
+              webServer({
+                  command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port}',
+                  port: ${port},
+                  setBaseURL: false,
+              }),
+              webServer({
+                  command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port + 1}',
+                  port: ${port + 1},
+              }),
+             ],
+            };
+          `,
+    }, undefined, { DEBUG: 'pw:webserver' });
+    expect(result.exitCode).toBe(0);
+    expect(result.passed).toBe(1);
+  });
+
+  test('plugins takes precedence over legacy config', async ({ runInlineTest }, { workerIndex }) => {
+    const port = workerIndex + 10500;
+    const result = await runInlineTest({
+      'test.spec.ts': `
+          const { test } = pwt;
+          test('connect to the server', async ({baseURL, page}) => {
+            expect(baseURL).toBe('http://localhost:${port + 1}');
+          });
+        `,
+      'playwright.config.ts': `
+          import { webServer } from '@playwright/test/lib/plugins';
+          module.exports = {
+            plugins: [
+              webServer({
+                  command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port + 1}',
+                  port: ${port + 1},
+              }),
+            ],
+            webServer: {
+              command: 'node ${JSON.stringify(SIMPLE_SERVER_PATH)} ${port}',
+              port: ${port},
+            }
+          };
+          `,
+    }, undefined, { DEBUG: 'pw:webserver' });
+    expect(result.exitCode).toBe(0);
+    expect(result.passed).toBe(1);
+  });
+});
